tests/wpt/tests/html/canvas/offscreen/canvas-host/2d.canvas.host.type.delete.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.canvas.host.type.delete</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.host.type.delete</h1>
<p class="desc">OffscreenCanvas interface object is [[Configurable]]</p>


<script>
var t = async_test("OffscreenCanvas interface object is [[Configurable]]");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  _assertSame(delete window.OffscreenCanvas, true, "delete window.OffscreenCanvas", "true");
  _assertSame(window.OffscreenCanvas, undefined, "window.OffscreenCanvas", "undefined");
  t.done();

});
</script>
